+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 1.1.6 - In progress
 
-- [Feature] Compose Theme(support dark theme, but not finished)
+- [Feature] Compose Theme(with support switch theme)
 - [Feature] Add black flipper support (update and info screen)
 - [Feature] Compose Theme(module `core:ui:theme`, module `archive`)
 - [Feature] Auto disconnect after 5 minutes inactivity